Remove reconnect on connection close.
I'm doing several stress test and the only case where this seems to
occur is not when the connection is in a "bug" state.

Basically I found that this happens (correctly) when we are using
multiple threads and we `stop` the connection.

So maybe on thread close the connection meanwhile the other is receiving
an `_heartbeat_`.

So I'm going to remove this also because the `reconnection` in this
scenario should not happen and so the `write` raise an unhandled
exception.